fix: route structure
